+++ src/main/java/com/example/yeobee/common/entity/BaseEntity.java
@@ -1,12 +1,9 @@
 package com.example.yeobee.common.entity;
 
 import jakarta.persistence.*;
-import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
 
 import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Getter
