+++ src/main/java/com/example/yeobee/core/user/presentation/UserController.java
@@ -28,10 +28,4 @@ public class UserController {
         @RequestBody UserUpdateRequestDto request) {
         return ResponseEntity.ok(userService.updateUserInfo(user, request));
     }
-
-    @DeleteMapping(value = "/me")
-    public ResponseEntity<Void> deleteUser(@AuthUser User user) {
-        userService.deleteUser(user);
-        return ResponseEntity.noContent().build();
-    }
 }
